added support for msys2 mingwtoolchains and fix x64 cocos/audio on windows platform for MSVC and Mingw. Also fixed a few very bad type casts very important for future Arm64 bit processors

+++ cocos/2d/CCSprite.cpp
@@ -613,7 +613,7 @@ void Sprite::draw(void)
     long offset = 0;
     setGLBufferData(&_quad, 4 * kQuadSize, 0);
 #else
-    long offset = (long)&_quad;
+    size_t offset = (size_t)&_quad;
 #endif // EMSCRIPTEN
 
     // vertex
